restProxy tests, memoryProxy core-shell architecture, tests
